requirements.txt
alembic==1.7.7
blinker==1.4
click==8.0.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==1.0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
python-decouple==3.6
SQLAlchemy==1.4.32
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
